Initial version
A simple interface for defining MPD commands, closely modelled
after Network.MPD.Commands.Applicative, as well as the machinery required
to run those commands.

The API eschews flexibility for simplicity, and is limited to defining
and running commands, with few possibilities for hooking into this process,
other than to define folds for individual responses.
There is no connection persistence, the implementation re-connects to MPD
for every run.

All response processing is defined on lazy bytestring, decoupling
it entirely from the response reading step.

The pure response handling code appears to run in constant time and space even
in perverse applications, e.g.

   replicateM_ 100000 (run listallinfo)

uses constant memory and CPU time on my machine for the duration of the run.