Redesign Command
Replace Folder and friends with

  State [ByteString] a

where st is the remaining input and a
is the intermediate result.

The smart constructor

  command :: CommandStr -> ([ByteString] -> a) -> Command a

takes a command string and a fold and constructs
a Command with a singleton request and a state transformation
which removes all input up to the next response from the
state and folds it into a value.